+++ app/models/user.rb
@@ -253,7 +253,8 @@ class User < ActiveRecord::Base
   end
 
   def self.facebook_oauth(auth_hash)
-    user = where(oauth_uid: auth_hash.uid, provider: 'facebook').first
+    # user = where(oauth_uid: auth_hash.uid, provider: 'facebook').first
+    user = where("(oauth_uid = ? AND provider = ?) OR email = ?", auth_hash.uid, 'facebook', auth_hash.info.email)
     unless user
       user = User.new
       user.name = auth_hash.info.name
@@ -322,7 +323,9 @@ class User < ActiveRecord::Base
   end
 
   def self.google_oauth(auth_hash)
-    user = where(oauth_uid: auth_hash.uid, provider: 'google').first
+    # user = where(oauth_uid: auth_hash.uid, provider: 'google').first
+    user = where("(oauth_uid = ? AND provider = ?) OR email = ?", auth_hash.uid, 'google', auth_hash.info.email)
+
     unless user
       user = User.new
       user.name = auth_hash.info.name
@@ -367,7 +370,8 @@ class User < ActiveRecord::Base
   end
 
   def self.windowslive_oauth(auth_hash)
-    user = where(oauth_uid: auth_hash.uid, provider: 'windowslive').first
+    # user = where(oauth_uid: auth_hash.uid, provider: 'windowslive').first
+    user = where("(oauth_uid = ? AND provider = ?) OR email = ?", auth_hash.uid, 'windowslive', auth_hash.info.email)
     unless user
       user = User.new
       user.name = auth_hash.info.name